Initial Upload
All functionality of the scraper appears to be working quite well. I’m
not super thrilled with how the dates are written into the file name
but it should pass mustard. I could utilize another NPM package to fix
this but will wait for feedback before proceeding.

The one thing I haven’t checked for is if errors are logged
appropriately to the log file. Will have to run through a number of
scenarios where the file should fail and see what the outcomes are.